Showing existing Firebases in login form
Please pardon me, this is the first time I'm using ReactJS, so there's probably a whole lot that can be done better. Oh, and it's also my first Chrome extension.

What this does is it fetches your `adminToken` cookie from `*.firebase.com` and uses it to get a list of your Firebases. These then are shown under the login form.